% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeETLjob.R
\name{makeETLjob}
\alias{makeETLjob}
\title{Start an etl job by creating a directory with supporting documents}
\usage{
makeETLjob(location = ".", name)
}
\arguments{
\item{location}{Location where the new directory will be}

\item{name}{The name of the new directory}
}
\description{
Start an etl job by creating a directory with supporting documents
}
\details{
This function will create a new directory with the following CSV files
    in it: sources, relationships, transformations, codes, summarize, reshape, and load.
}
\examples{
makeETLjob(name = "my_new_job")
}

