% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_etl_job.R
\name{run_etl_job}
\alias{run_etl_job}
\title{Run an etl job}
\usage{
run_etl_job(job_location)
}
\arguments{
\item{location}{Location of job directory}
}
\description{
Run an etl job
}
\examples{
run_etl_job("my_new_job")
}

