% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_odbc_credential.R
\name{set_odbc_credential}
\alias{set_odbc_credential}
\alias{unset_odbc_credential}
\title{Set credentials for an ODBC data source}
\usage{
set_odbc_credential(dsn, usr, pwd)

unset_odbc_credential(dsn)
}
\arguments{
\item{dsn}{The name of the data source name}

\item{usr}{The user name for the connection}

\item{pwd}{The password for the connection}
}
\description{
Set credentials for an ODBC data source
}
\section{Functions}{
\itemize{
\item \code{unset_odbc_credential}: Unset the ODBC credentials
}}
\examples{
set_odbc_credential("whiskey", "foo", "bar")
Sys.getenv(c("WHISKEY_USR", "WHISKEY_PWD"))
unset_odbc_credential("whiskey")
Sys.getenv(c("WHISKEY_USR", "WHISKEY_PWD"), unset = NA)
}

